Fixed Player Delete
Fixed a minor issue with the player delete code. Now it will remove the player entirely from the array although it will not delete the object.
The last code only deleted the first variable of the object which was the name.

I also made a few things look a bit nicer for the user.